Update mail notify setting controller for fixing mail notify setting empty error after click back button when edit

+++ marquee-web-main/app/controllers/admin/mail_notify_settings_controller.rb
@@ -8,8 +8,8 @@ class Admin::MailNotifySettingsController < InheritedResources::Base
   def create
     project_id = params[:project_id]
     mail = params[:mail_notify_setting][:mail]
-    mail_notify_group_ids = params[:mail_notify_setting][:mail_notify_group_ids]
-    test_type_ids = params[:mail_notify_setting][:test_type_ids]
+    mail_notify_group_ids = params[:mail_notify_setting][:mail_notify_group_ids].delete_if {|mng_id| mng_id == "" }
+    test_type_ids = params[:mail_notify_setting][:test_type_ids].delete_if {|tt_id| tt_id == "" }
 
     existing_mail_notify_settings_list = MailNotifySetting.joins(:mail_notify_groups, :test_types).where(:project_id => project_id, :mail => mail, "mail_notify_groups.id" => mail_notify_group_ids, "test_types.id" => test_type_ids)
 
@@ -20,14 +20,12 @@ class Admin::MailNotifySettingsController < InheritedResources::Base
         mns.project = project
 
         mail_notify_group_ids.each do |mng_id|
-          mng = nil
-          mng = MailNotifyGroup.find(mng_id) unless mng_id == ""
+          mng = MailNotifyGroup.find(mng_id)
           mns.mail_notify_groups << mng if mng
         end
 
         test_type_ids.each do |type_id|
-          test_type = nil
-          test_type = TestType.find(type_id) unless type_id == ""
+          test_type = TestType.find(type_id)
           mns.test_types << test_type if test_type
         end
         mns.save
@@ -44,29 +42,31 @@ class Admin::MailNotifySettingsController < InheritedResources::Base
     id = params[:id]
     project_id = params[:project_id]
     mail = params[:mail_notify_setting][:mail]
-    mail_notify_group_ids = params[:mail_notify_setting][:mail_notify_group_ids]
-    test_type_ids = params[:mail_notify_setting][:test_type_ids]
-
-    existing_mail_notify_settings_list = MailNotifySetting.joins(:mail_notify_groups, :test_types).where(:project_id => project_id, :mail => mail, "mail_notify_groups.id" => mail_notify_group_ids, "test_types.id" => test_type_ids)
+    mail_notify_group_ids = params[:mail_notify_setting][:mail_notify_group_ids].delete_if {|mng_id| mng_id == "" }
+    test_type_ids = params[:mail_notify_setting][:test_type_ids].delete_if {|tt_id| tt_id == "" }
+    
+    existing_mail_notify_settings_list = MailNotifySetting.joins(:mail_notify_groups, :test_types).where('mail_notify_settings.id != ?', id).where(:project_id => project_id, :mail => mail, "mail_notify_groups.id" => mail_notify_group_ids, "test_types.id" => test_type_ids)
 
     mns = MailNotifySetting.find(id)
     respond_to do |format|
       if mns
         mns.mail = mail
-        mns.mail_notify_groups.delete_all
-        mns.test_types.delete_all
         
         if existing_mail_notify_settings_list && existing_mail_notify_settings_list.empty?
-          mail_notify_group_ids.each do |mng_id|
-            m = nil
-            m = MailNotifyGroup.find(mng_id) unless mng_id == ""
-            mns.mail_notify_groups << m if m
+          if not mail_notify_group_ids.empty?
+            mns.mail_notify_groups.delete_all
+            mail_notify_group_ids.each do |mng_id|
+              m = MailNotifyGroup.find(mng_id)
+              mns.mail_notify_groups << m if m
+            end
           end
 
-          test_type_ids.each do |type_id|
-            test_type = nil
-            test_type = TestType.find(type_id) unless type_id == ""
-            mns.test_types << test_type if test_type
+          if not mail_notify_group_ids.empty?
+            mns.test_types.delete_all
+            test_type_ids.each do |type_id|
+              test_type = TestType.find(type_id)
+              mns.test_types << test_type if test_type
+            end
           end
           mns.save
           format.js { render :json => {:result => "success"}}
